+++ dpnp/backend/CMakeLists.txt
@@ -186,7 +186,9 @@ set(DPNP_SRC
     kernels/dpnp_krnl_statistics.cpp
     src/dpnp_iface_fptr.cpp
     src/memory_sycl.cpp
-    src/queue_sycl.cpp)
+    src/queue_sycl.cpp
+    src/verbose.cpp
+    )
 
 if(DPNP_STATIC_LIB_ENABLE)
   add_library(dpnp_backend_c STATIC ${DPNP_SRC})
