+++ dpnp/backend/src/queue_sycl.hpp
@@ -46,6 +46,8 @@
 
 #include "dpnp_pstl.hpp" // this header must be included after <mkl.hpp>
 
+#include "verbose.hpp"
+
 namespace mkl_rng = oneapi::mkl::rng;
 
 #define DPNP_QUEUE      backend_sycl::get_queue()
@@ -145,9 +147,5 @@ public:
         return *rng_engine;
     }
 };
-
-bool is_verbose_mode();
-void set_barrier_event(cl::sycl::queue queue, sycl::vector_class<sycl::event> & depends);
-void verbose_print(std::string header, cl::sycl::event first_event, cl::sycl::event last_event);
 
 #endif // QUEUE_SYCL_H
